fix: menu under cog icon appears offset to the right
Vuetify expects to be alone on the page with a top level v-app
component. Since this is not the case in zmast.html we use a
modified version of Vuetify that works around this. This is the
same package used in ipyvuetify in the notebook, since that
environment has the same issue.

This also prevents a lot of style conflicts between Vuetify and
Bootstrap.